view folder renamed to resources client view classes moved to view
