+++ client/resources/Bombsprite.png
