+++ client/resources/Firesprite.png
