+++ client/resources/Ringsprite.png
