+++ client/resources/Sparksprite.png
