+++ client/resources/background.png
