+++ client/resources/example.glade
