+++ client/resources/insert_name.glade
