+++ client/resources/main_window.glade
