+++ client/resources/name_entry.glade
