+++ client/resources/stage_pick.glade
