+++ client/resources/stage_pick_box.glade
