+++ client/resources/welcome_screen.glade
