+++ client/resources/welcome_screen_bg.glade
