+++ client/view/GladeLoader.cpp
