+++ client/view/GladeLoader.h
